Move UT to DataTypeSuite
